// src/shared/api/generated/ipc.ts
export type OrderedFieldDto = { enabled: boolean, order: number, name: string, value: string, };

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BodyFilePathDto = { "type": "RELATIVE", path: string, } | { "type": "ABSOLUTE", path: string, };

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BodyFileReferenceDto = { path: BodyFilePathDto, fileName: string, size: bigint, modifiedAtEpochSeconds: bigint | null, sha256: string, };
// ...
export type MultipartPartDto = { "type": "FIELD", enabled: boolean, order: number, name: string, value: string, } | { "type": "FILE", enabled: boolean, order: number, name: string, file: BodyFileReferenceDto, };

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RequestBodyDto = { "type": "NONE" } | { "type": "RAW", content: string, } | { "type": "URL_ENCODED", fields: Array<OrderedFieldDto>, } | { "type": "MULTIPART", parts: Array<MultipartPartDto>, } | { "type": "BINARY", file: BodyFileReferenceDto, };

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RequestContentDto = { name: string, method: string, url: string, body: RequestBodyDto, query: Array<OrderedFieldDto>, headers: Array<OrderedFieldDto>, };
// ...
export type RequestDraftDto = { id: string, workspaceId: string, savedRequestId: string | null, content: RequestContentDto, isDirty: boolean, };